Move Goa extensions to `api/server/common`
