+++ api/templates/design.go
@@ -9,13 +9,13 @@ import (
 	"goa.design/goa/v3/http/codegen/openapi"
 	cors "goa.design/plugins/v3/cors/dsl"
 
-	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/templates/extension/anytype"
-	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/templates/extension/dependencies"
-	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/templates/extension/genericerror"
-	"github.com/keboola/keboola-as-code/internal/pkg/api/server/templates/extension/oneof"
-	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/templates/extension/oneof"
-	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/templates/extension/operationid"
-	. "github.com/keboola/keboola-as-code/internal/pkg/api/server/templates/extension/token"
+	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/common/extension/anytype"
+	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/common/extension/dependencies"
+	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/common/extension/genericerror"
+	"github.com/keboola/keboola-as-code/internal/pkg/api/server/common/extension/oneof"
+	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/common/extension/oneof"
+	_ "github.com/keboola/keboola-as-code/internal/pkg/api/server/common/extension/operationid"
+	. "github.com/keboola/keboola-as-code/internal/pkg/api/server/common/extension/token"
 	"github.com/keboola/keboola-as-code/internal/pkg/encoding/json"
 )
