+++ internal/pkg/api/server/common/extension/oneof/oneof.go
